requirements.txt
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
anthropic==0.28.0
